Move BuildLogger to root package
