+++ jib-core/src/main/java/com/google/cloud/tools/jib/BuildLogger.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.builder;
+package com.google.cloud.tools.jib;
 
 public interface BuildLogger {
