+++ jib-core/src/main/java/com/google/cloud/tools/jib/frontend/MainClassFinder.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.jib.frontend;
 
-import com.google.cloud.tools.jib.builder.BuildLogger;
+import com.google.cloud.tools.jib.BuildLogger;
 import com.google.cloud.tools.jib.filesystem.DirectoryWalker;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
