+++ jib-core/src/test/java/com/google/cloud/tools/jib/builder/TestBuildLogger.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.tools.jib.builder;
 
+import com.google.cloud.tools.jib.BuildLogger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
